+++ src/shared_base/core/asDataPredictorArchiveNcepReanalysis2.cpp
@@ -299,7 +299,7 @@ bool asDataPredictorArchiveNcepReanalysis2::ExtractFromFiles(asGeoAreaCompositeG
         VVectorShort vectData;
         VVectorShort vectData360;
 
-        for (int i_area = 0; i_area<compositeData.size(); i_area++)
+        for (int i_area = 0; i_area<(int)compositeData.size(); i_area++)
         {
             // Check if necessary to load the data of lon=360 (so lon=0)
             bool load360 = false;
@@ -321,7 +321,7 @@ bool asDataPredictorArchiveNcepReanalysis2::ExtractFromFiles(asGeoAreaCompositeG
                 {
                     // Correction if the lon 360 degrees is required (doesn't exist)
                     load360 = true;
-                    for (int i_check = 0; i_check<compositeData.size(); i_check++)
+                    for (int i_check = 0; i_check<(int)compositeData.size(); i_check++)
                     {
                         // If so, already loaded in another composite
                         if(dataArea->GetComposite(i_check).GetXmin() == 0)
@@ -537,7 +537,7 @@ bool asDataPredictorArchiveNcepReanalysis2::ExtractFromFiles(asGeoAreaCompositeG
         if (compositeData[0].capacity()==0)
         {
             int totSize = 0;
-            for (int i_area = 0; i_area<compositeData.size(); i_area++)
+            for (int i_area = 0; i_area<(int)compositeData.size(); i_area++)
             {
                 int indexLengthLat = vectIndexLengthLat[i_area];
                 int indexLengthLon = vectIndexLengthLon[i_area];
@@ -547,7 +547,7 @@ bool asDataPredictorArchiveNcepReanalysis2::ExtractFromFiles(asGeoAreaCompositeG
         }
 
         // Transfer data
-        for (int i_area = 0; i_area<compositeData.size(); i_area++)
+        for (int i_area = 0; i_area<(int)compositeData.size(); i_area++)
         {
             // Extract data
             int indexLengthLat = vectIndexLengthLat[i_area];
